# openapi_agent_workflow_platform_mvp_v1.yaml
openapi: 3.0.3
info:
  title: Agent Workflow Platform MVP API
  version: 0.1.0
  description: MVP API draft for workflow authoring, publishing, running, tracing, knowledge bases, tools, models, and secrets.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /knowledge-bases:
    post:
      tags: [Knowledge Bases]
      summary: Create knowledge base
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateKnowledgeBaseRequest"
      responses:
        "200":
          description: Knowledge base
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        default:
          $ref: "#/components/responses/ErrorResponse"
    get:
      tags: [Knowledge Bases]
      summary: List knowledge bases
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: keyword
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Knowledge base list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBaseListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /knowledge-bases/{kb_id}:
    get:
      tags: [Knowledge Bases]
      summary: Get knowledge base detail
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      responses:
        "200":
          description: Knowledge base
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KnowledgeBase"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /knowledge-bases/{kb_id}/documents:
    post:
      tags: [Documents]
      summary: Upload document to knowledge base
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                metadata_json:
                  type: string
      responses:
        "200":
          description: Uploaded document
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadDocumentResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
    get:
      tags: [Documents]
      summary: List documents in knowledge base
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/DocumentStatus"
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /knowledge-bases/{kb_id}/retrieve:
    post:
      tags: [Knowledge Bases]
      summary: Test knowledge retrieval
      parameters:
        - $ref: "#/components/parameters/KnowledgeBaseId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetrieveKnowledgeRequest"
      responses:
        "200":
          description: Retrieved chunks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetrieveKnowledgeResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /documents/{document_id}:
    get:
      tags: [Documents]
      summary: Get document detail
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Document detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        default:
          $ref: "#/components/responses/ErrorResponse"
    delete:
      tags: [Documents]
      summary: Soft delete document
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: Delete result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
  /documents/{document_id}/retry:
    post:
      tags: [Documents]
      summary: Retry document processing
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                from_stage:
                  type: string
                  example: embedding
      responses:
        "200":
          description: Retry result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UploadDocumentResponse"
        default:
          $ref: "#/components/responses/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    KnowledgeBaseId:
      name: kb_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
    DocumentId:
      name: document_id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    ErrorResponse:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    JsonObject:
      type: object
      additionalProperties: true
    Timestamp:
      type: string
      format: date-time
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              $ref: "#/components/schemas/JsonObject"
        request_id:
          type: string
...
    KnowledgeBase:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
        status:
          type: string
        embedding_model:
          type: string
        config_json:
          $ref: "#/components/schemas/JsonObject"
        created_at:
          $ref: "#/components/schemas/Timestamp"
        updated_at:
          $ref: "#/components/schemas/Timestamp"
    CreateKnowledgeBaseRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        embedding_model:
          type: string
        config:
          $ref: "#/components/schemas/JsonObject"
    KnowledgeBaseListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeBase"
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    DocumentStatus:
      type: string
      enum: [uploaded, parsing, chunking, embedding, indexed, failed, deleted]
    Document:
      type: object
      properties:
        id:
          type: integer
          format: int64
        knowledge_base_id:
          type: integer
          format: int64
        file_name:
          type: string
        file_type:
          type: string
        file_size:
          type: integer
          format: int64
        status:
          $ref: "#/components/schemas/DocumentStatus"
        error_stage:
          type: string
          nullable: true
        error_message:
          type: string
          nullable: true
        metadata_json:
          $ref: "#/components/schemas/JsonObject"
        created_at:
          $ref: "#/components/schemas/Timestamp"
    UploadDocumentResponse:
      type: object
      properties:
        document_id:
          type: integer
          format: int64
        knowledge_base_id:
          type: integer
          format: int64
        file_name:
          type: string
        status:
          $ref: "#/components/schemas/DocumentStatus"
        processing_job_id:
          type: integer
          format: int64
    DocumentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Document"
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
    RetrieveKnowledgeRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
        top_k:
          type: integer
          default: 5
        score_threshold:
          type: number
          format: float
          default: 0.65
    KnowledgeChunkResult:
      type: object
      properties:
        chunk_id:
          type: string
        content:
          type: string
        score:
          type: number
        source:
          $ref: "#/components/schemas/JsonObject"
    RetrieveKnowledgeResponse:
      type: object
      properties:
        chunks:
          type: array
          items:
            $ref: "#/components/schemas/KnowledgeChunkResult"
...